requirements.txt
discord.py>=2.3,<3
yfinance>=0.2.31
pandas>=2.0
ta>=0.11
python-dotenv>=1.0
aiosqlite>=0.19
tavily-python>=0.3.0
pytest>=7.0
pytest-cov>=4.0
